+++ src/main/java/com/perimeterx/http/RequestWrapper.java
@@ -1,5 +1,7 @@
 package com.perimeterx.http;
 
+import com.perimeterx.utils.PXLogger;
+
 import javax.servlet.ServletInputStream;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletRequestWrapper;
@@ -16,26 +18,18 @@ import java.util.Map;
  * This enables reading the request body multiple times
  * **/
 public class RequestWrapper extends HttpServletRequestWrapper {
-
+    private final PXLogger logger = PXLogger.getLogger(RequestWrapper.class);
     private String body;
     private final Map<String, String> customHeaders;
 
-    public RequestWrapper(HttpServletRequest request) throws IOException {
+    public RequestWrapper(HttpServletRequest request) {
         super(request);
         this.customHeaders = new HashMap<>();
-        this.body = "";
-
-        final BufferedReader bufferedReader = request.getReader();
-        String line;
-
-        while ((line = bufferedReader.readLine()) != null){
-            this.body += line;
-        }
     }
 
     @Override
-    public ServletInputStream getInputStream() {
-        final ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(body.getBytes());
+    public ServletInputStream getInputStream() throws IOException {
+        final ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(getBody().getBytes());
         return new ServletInputStream() {
             public int read() {
                 return byteArrayInputStream.read();
@@ -62,7 +56,18 @@ public class RequestWrapper extends HttpServletRequestWrapper {
         this.customHeaders.put(name, value);
     }
 
-    public String getBody() {
+    public synchronized String getBody() throws IOException {
+        if(body == null) {
+            this.body = "";
+            char[] buffer = new char[4096];
+            final BufferedReader reader = this.getRequest().getReader();
+            StringBuilder builder = new StringBuilder();
+            int numChars;
+            while ((numChars = reader.read(buffer)) >= 0) {
+                builder.append(buffer, 0, numChars);
+            }
+            body = builder.toString();
+        }
         return body;
     }
 }