+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [v6.5.0](https://github.com/PerimeterX/perimeterx-java-sdk/compare/v6.5.0...HEAD) (2023-03-04)
+- Adding custom is sensitive configuration option
+- Lazy read the request body
+- Added new custom parameters function signature which receives the original HTTP request
+- Reading the body binary instead of textually
+
 ## [v6.4.5](https://github.com/PerimeterX/perimeterx-java-sdk/compare/v6.4.5...HEAD) (2023-01-11)
 - Fixed invalid http connections for risk requests bug.